requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cors==1.0.1
dateparser==1.1.1
Deprecated==1.2.13
filelock==3.6.0
Flask==2.1.1
Flask-Cors==3.0.10
frozenlist==1.3.0
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
pycparser==2.21
pyparsing==3.0.8
PySocks==1.7.1
python-binance==1.0.16
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
redis==4.2.2
regex==2022.3.2
rel==0.4.7
requests==2.27.1
requests-file==1.5.1
six==1.16.0
tldextract==3.2.1
tzdata==2022.1
tzlocal==4.2
ujson==5.2.0
urllib3==1.26.9
websocket-client==1.3.2
websockets==10.2
Werkzeug==2.1.1
wrapt==1.14.0
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
gunicorn==20.1.0
